Created A scratch folder and added my scratch
This is some independent stuff that I did before I discovered the new excel format. This isn't too helpful aside from some of the syntax that I use here and some of the ideas that I used to make different weighted categories. This is using the data from msugrades.com when you download the master.xlsx file.